rename OnyxAlertDialog to OnyxAlertModal
