+++ packages/sit-onyx/src/components/OnyxAlertModal/examples/DefaultExample.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { iconCircleAttention } from "@sit-onyx/icons";
 import { ref } from "vue";
-import { OnyxAlertDialog, OnyxButton } from "../../../index.js";
+import { OnyxAlertModal, OnyxButton } from "../../../index.js";
 
 const isOpen = ref(false);
 </script>
@@ -10,7 +10,7 @@ const isOpen = ref(false);
   <div>
     <OnyxButton label="Show alert modal" @click="isOpen = true" />
 
-    <OnyxAlertDialog
+    <OnyxAlertModal
       :icon="{ icon: iconCircleAttention, color: 'danger' }"
       label="Confirm deletion"
       :open="isOpen"
@@ -22,6 +22,6 @@ const isOpen = ref(false);
         <OnyxButton label="Cancel" color="neutral" mode="plain" autofocus @click="isOpen = false" />
         <OnyxButton label="Delete" color="danger" />
       </template>
-    </OnyxAlertDialog>
+    </OnyxAlertModal>
   </div>
 </template>
