+++ packages/sit-onyx/src/components/OnyxAlertModal/types.ts
@@ -1,7 +1,7 @@
 import type { OnyxDialogProps } from "../OnyxDialog/types.js";
 import type { OnyxIconProps } from "../OnyxIcon/types.js";
 
-export type OnyxAlertDialogProps = Omit<OnyxDialogProps, "modal" | "alert" | "alignment"> & {
+export type OnyxAlertModalProps = Omit<OnyxDialogProps, "modal" | "alert" | "alignment"> & {
   /**
    * Optional icon to show.
    */
